+++ android/src/main/kotlin/dev/msfjarvis/claw/android/ui/decorations/ClawNavigationBar.kt
@@ -26,15 +26,12 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.unit.dp
-import androidx.navigation.NavController
-import androidx.navigation.compose.currentBackStackEntryAsState
+import androidx.navigation3.runtime.NavKey
 import dev.chrisbanes.haze.HazeDefaults
 import dev.chrisbanes.haze.HazeState
 import dev.chrisbanes.haze.HazeStyle
 import dev.chrisbanes.haze.hazeEffect
 import dev.msfjarvis.claw.android.ui.navigation.AppDestinations
-import dev.msfjarvis.claw.android.ui.navigation.Destination
-import dev.msfjarvis.claw.android.ui.navigation.matches
 import dev.msfjarvis.claw.common.ui.FloatingNavigationBar
 import kotlinx.collections.immutable.ImmutableList
 
@@ -42,8 +39,9 @@ const val AnimationDuration = 100
 
 @Composable
 fun ClawNavigationBar(
-  navController: NavController,
   items: ImmutableList<NavigationItem>,
+  currentNavKey: NavKey?,
+  navigateTo: (NavKey) -> Unit,
   isVisible: Boolean,
   hazeState: HazeState,
   modifier: Modifier = Modifier,
@@ -85,10 +83,8 @@ fun ClawNavigationBar(
         containerColor =
           if (HazeDefaults.blurEnabled()) Color.Transparent else MaterialTheme.colorScheme.surface,
       ) {
-        val navBackStackEntry = navController.currentBackStackEntryAsState().value
-        val currentDestination = navBackStackEntry?.destination
         items.forEach { navItem ->
-          val isSelected = currentDestination.matches(navItem.destination)
+          val isSelected = currentNavKey == navItem.navKey
           NavigationBarItem(
             icon = {
               Crossfade(isSelected, label = "nav-label") {
@@ -104,11 +100,7 @@ fun ClawNavigationBar(
               if (isSelected) {
                 navItem.listStateResetCallback()
               } else {
-                navController.navigate(navItem.destination) {
-                  popUpTo(navController.graph.startDestinationId) { saveState = true }
-                  launchSingleTop = true
-                  restoreState = true
-                }
+                navigateTo(navItem.navKey)
               }
             },
             modifier = Modifier.testTag(navItem.label.uppercase()),
@@ -123,7 +115,7 @@ class NavigationItem
 private constructor(
   val icon: ImageVector,
   val label: String,
-  val destination: Destination,
+  val navKey: NavKey,
   val selectedIcon: ImageVector,
   val listStateResetCallback: () -> Unit,
 ) {
@@ -133,7 +125,7 @@ private constructor(
   ) : this(
     destination.icon,
     destination.label,
-    destination.destination,
+    destination.navKey,
     destination.selectedIcon,
     listStateResetCallback,
   ) {}